feat: Delete Person by clicking on Trash Icon
Added a button with trash emoji inside of it. This button is rendered in
the Header component of all the cards, and if clicked, will delete the
card from the server.
For now, we are re-fetching all the content from the server to trigger a
re-render. That is, the `Get all Persons` action is being dispatched on
successful deletion (and addition).
However, this somewhat defeats the purpose of react, and will
be fixed later on.