feat: show confirm dialog before deletion
added a new key "deletePopup" while sending the state back from the
inital get request. Using this key for showing/hiding a confirm dialog
when user clicks on the delete button.
Created types, actions, and reducers corresponding to SHOW_POPUP &
HIDE_POPUP functionality.
This was required because accidental clicks would remove the card
instantly, which is not desired behavior.

Changed the earlier implementation for showing updated states
(introduced forced_rendering) as the component was not being re-rendered
even though the state was changing. This was more of a POC approach and
should be fixed later on, where the structure of data remains flat.